Move inventoryStateWrapper to test helpers
