+++ app/controllers/foreman_rh_cloud/react_controller.rb
@@ -1,6 +1,6 @@
 module ForemanRhCloud
   class ReactController < ::ApplicationController
-    layout "foreman_rh_cloud/application" 
+    layout "foreman_rh_cloud/application"
 
     def inventory_upload
     end
