+++ webpack/ForemanRhCloudTestHelpers.js
@@ -1,0 +1,5 @@
+export const inventoryStateWrapper = innerState => ({
+  ForemanRhCloud: {
+    inventoryUpload: { ...innerState },
+  },
+});
